requirements-test.txt
# Development dependencies

pytest
coverage
pyright
PyHamcrest
pytest-asyncio
httpx
playwright
pytest-playwright